build_wordvec : draft process logic
